Isolate version types
