+++ src/models/version/raw-version.ts
@@ -1,0 +1,17 @@
+/*************************************************************************
+ * Copyright 2020 Gravwell, Inc. All rights reserved.
+ * Contact: <legal@gravwell.io>
+ *
+ * This software may be modified and distributed under the terms of the
+ * MIT license. See the LICENSE file for details.
+ **************************************************************************/
+
+/**
+ * @examples
+ * ```ts
+ * '3.5.7'
+ * '3'
+ * '1.5'
+ * ```
+ */
+export type RawVersion = string;
