+++ src/models/version/to-version.ts
@@ -6,38 +6,10 @@
  * MIT license. See the LICENSE file for details.
  **************************************************************************/
 
-import { isInteger, isNumber, isString } from 'lodash';
-
-/**
- * @examples
- * ```ts
- * '3.5.7'
- * '3'
- * '1.5'
- * ```
- */
-export type RawVersion = string;
-
-export type RawVersionObject = {
-	Major: number;
-	Minor?: number;
-	Point?: number;
-};
-
-export interface Version {
-	major: number;
-	minor: number;
-	patch: number;
-}
-
-export const isVersion = (value: any): value is Version => {
-	try {
-		const v = <Version>value;
-		return isInteger(v.major) && isInteger(v.minor) && isInteger(v.patch);
-	} catch {
-		return false;
-	}
-};
+import { isNumber, isString } from 'lodash';
+import { RawVersion } from './raw-version';
+import { RawVersionObject } from './raw-version-object';
+import { Version } from './version';
 
 export const toVersion = (raw: RawVersion | RawVersionObject | number): Version => {
 	if (isString(raw) || isNumber(raw)) {
