+++ src/models/version/version.ts
@@ -1,0 +1,13 @@
+/*************************************************************************
+ * Copyright 2020 Gravwell, Inc. All rights reserved.
+ * Contact: <legal@gravwell.io>
+ *
+ * This software may be modified and distributed under the terms of the
+ * MIT license. See the LICENSE file for details.
+ **************************************************************************/
+
+export interface Version {
+	major: number;
+	minor: number;
+	patch: number;
+}
